v0.0.1
Changed the package name to obtain harmony among all files.
Added changelog.
Added feedback toast after submiting barcode to server. (Actually barcode is NOT sent to server. The sending process is randomly percieved as succesful or unsuccesful)
EditText has now hint and is longer in order to show all of the hint.